feat: hide elements based on permissions; closes #55

+++ lib/features/channel/ui/sheets/channel_list_bottomsheet.dart
@@ -2,6 +2,7 @@ import 'package:Self.Tube/common/ui/widgets/containers/list_section_container.da
 import 'package:Self.Tube/common/ui/widgets/dialogs/confirmation_dialog.dart';
 import 'package:Self.Tube/common/ui/widgets/sheets/bottomsheet_template.dart';
 import 'package:Self.Tube/features/channel/data/api/channel_api.dart';
+import 'package:Self.Tube/features/onboarding/domain/user_session.dart';
 import 'package:flutter/material.dart';
 import 'package:Self.Tube/l10n/generated/app_localizations.dart';
 import 'package:share_plus/share_plus.dart';
@@ -27,19 +28,20 @@ Future<void> showChannelListBottomSheet({
               );
             },
           ),
-          ListTile(
-            leading: Icon(Icons.delete_forever),
-            title: Text(localizations.channelDelete),
-            onTap: () {
-              ConfirmationDialog(
-                context: context, 
-                onSure: () {
-                  ChannelApi().deleteChannel(channel.channelId);
-                  Navigator.pop(context);
-                }
-              );
-            },
-          ),
+          if (UserSession.isPrivileged)
+            ListTile(
+              leading: Icon(Icons.delete_forever),
+              title: Text(localizations.channelDelete),
+              onTap: () {
+                ConfirmationDialog(
+                  context: context, 
+                  onSure: () {
+                    ChannelApi().deleteChannel(channel.channelId);
+                    Navigator.pop(context);
+                  }
+                );
+              },
+            ),
         ]
       ),
     ]
